Rename account.rs to account_trie.rs
